[ui] Lazy-import markdown and react-dates
