+++ js_modules/dagster-ui/packages/ui-core/src/ui/Markdown.tsx
@@ -1,69 +1,8 @@
 import {FontFamily, colorTextLight} from '@dagster-io/ui-components';
 import * as React from 'react';
-import ReactMarkdown from 'react-markdown';
-import rehypeHighlight from 'rehype-highlight';
-import rehypeSanitize, {defaultSchema} from 'rehype-sanitize';
-import gfm from 'remark-gfm';
 import styled from 'styled-components';
 
-import 'highlight.js/styles/github.css';
-
-const sanitizeConfig = {
-  ...defaultSchema,
-  protocols: {
-    src: [...(defaultSchema.protocols?.src ?? []), 'data'],
-  },
-  attributes: {
-    ...defaultSchema.attributes,
-    span: [
-      ...(defaultSchema.attributes?.span || []),
-      // List of all allowed tokens:
-      [
-        'className',
-        'hljs-addition',
-        'hljs-attr',
-        'hljs-attribute',
-        'hljs-built_in',
-        'hljs-bullet',
-        'hljs-char',
-        'hljs-code',
-        'hljs-comment',
-        'hljs-deletion',
-        'hljs-doctag',
-        'hljs-emphasis',
-        'hljs-formula',
-        'hljs-keyword',
-        'hljs-link',
-        'hljs-literal',
-        'hljs-meta',
-        'hljs-name',
-        'hljs-number',
-        'hljs-operator',
-        'hljs-params',
-        'hljs-property',
-        'hljs-punctuation',
-        'hljs-quote',
-        'hljs-regexp',
-        'hljs-section',
-        'hljs-selector-attr',
-        'hljs-selector-class',
-        'hljs-selector-id',
-        'hljs-selector-pseudo',
-        'hljs-selector-tag',
-        'hljs-string',
-        'hljs-strong',
-        'hljs-subst',
-        'hljs-symbol',
-        'hljs-tag',
-        'hljs-template-tag',
-        'hljs-template-variable',
-        'hljs-title',
-        'hljs-type',
-        'hljs-variable',
-      ],
-    ],
-  },
-};
+const MarkdownWithPlugins = React.lazy(() => import('./MarkdownWithPlugins'));
 
 interface Props {
   children: string;
@@ -72,11 +11,9 @@ interface Props {
 export const Markdown = (props: Props) => {
   return (
     <Container>
-      <ReactMarkdown
-        {...props}
-        remarkPlugins={[gfm]}
-        rehypePlugins={[rehypeHighlight, [rehypeSanitize, sanitizeConfig]]}
-      />
+      <React.Suspense fallback={<div />}>
+        <MarkdownWithPlugins {...props} />
+      </React.Suspense>
     </Container>
   );
 };
